Change all image paths to .webp extension
The People page is considerably slow with mainstream image formats,
since they have less effective compression strategies. To resolve this,
I wrote a script that converted every image to the webp format and
changed all the image paths to use the webp extension. The new images
are updated on the S3 bucket.

I've kept the old images in the S3 bucket just in case.

Signed-off-by: Lucas Sta Maria <lucas@priime.dev>